requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-types==0.7.0
attrs==25.4.0
beautifulsoup4==4.14.3
cachetools==6.2.4
colorama==0.4.6
frozenlist==1.8.0
greenlet==3.3.1
idna==3.11
loguru==0.7.3
lxml==6.0.2
multidict==6.7.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.3
soupsieve==2.8.3
SQLAlchemy==2.0.46
typing-inspection==0.4.2
typing_extensions==4.15.0
win32_setctime==1.2.0
yarl==1.22.0
